==> 02_generate_dataset.ipynb <==
# Generate a synthetic dataset
Generate valid and invalid queries, including queries about unavailable hotels and irrelevant questions.

import random
from datetime import date, timedelta

import pandas as pd
from ollama import ChatResponse, chat
from tqdm import tqdm

%load_ext chime

def generate_random_date(start_date: date, end_date: date):
    """
    Generates a random date between two given datetime objects (inclusive).

    Args:
        start_date (datetime): The beginning of the date range.
        end_date (datetime): The end of the date range.

    Returns:
        datetime: A randomly generated date within the specified range.
    """
    # Calculate the difference in days between the start and end dates
    time_delta = end_date - start_date
    total_seconds = time_delta.total_seconds()

    # Generate a random number of seconds within that difference
    random_seconds = random.uniform(0, total_seconds)

    # Add the random seconds to the start date to get the random date
    random_date = start_date + timedelta(seconds=random_seconds)
    return random_date

def random_stay() -> tuple[date, date]:
    """Generate random check-in and check-out dates"""
    START_OF_PERIOD = date(2025, 8, 1)
    END_OF_PERIOD = date(2026, 8, 1)
    check_in = generate_random_date(START_OF_PERIOD, END_OF_PERIOD)
    check_out = generate_random_date(check_in, END_OF_PERIOD)
    return check_in, check_out

hotels_df = pd.read_csv('hotel_db.csv')

List of cities including some that are not in the hotels database

existing_cities = list(set(hotels_df.city.to_list()))
cities = existing_cities + ['Cape Town', 'Los Angeles', 'Keokuk', 'Paris']
cities

List of hotels including some that are not in the hotels database

hotels = list(set(hotels_df.name.to_list()))
hotels.extend(['Funny Inn', 'Nowhere Near', 'Keokuk Plaza', 'Golden Palace', 'Perfect Bliss'])
hotels

user_queries = []

def generate_query(prompt: str) -> str:
    system_message = 'Talk like a user interested in booking a hotel room. Your requests should be short. Do not ask for extra information'
    response: ChatResponse = chat(
        'mistral',
        messages=[
            {'role': 'system', 'content': system_message},
            {'role': 'user', 'content': prompt}
        ],
        stream=False,
        options={'temperature': 1.0}
    )
    return response.message.content

Generate valid availability queries:

for _ in tqdm(range(128)):
    city = random.choice(cities)
    check_in, check_out = random_stay()
    user_queries.append(
        {
            'available': city in existing_cities,
            'query': generate_query(f'Inquire about hotels available in {city} from {check_in} to {check_out}')
        }
    )

user_queries

Invalid queries with missing dates

%%chime
for _ in tqdm(range(64)):
    city = random.choice(cities)
    user_queries.append(
        {
            'available': city in existing_cities,
            'query': generate_query(f'Inquire about hotels available in {city}')
        }
    )

Invalid queries with missing checkout date

%%chime
for _ in tqdm(range(64)):
    city = random.choice(cities)
    check_in, check_out = random_stay()
    user_queries.append(
        {
            'available': city in existing_cities,
            'query': generate_query(f'Inquire about hotels available in {city} after {check_in}')
        }
    )

Valid booking requests

for _ in tqdm(range(128)):
    city = random.choice(cities)
    hotel = random.choice(hotels)
    check_in, check_out = random_stay()
    user_queries.append(
        {
            'available': len(hotels_df.query('city == @city & name == @hotel')) > 0,
            'query': generate_query(f'Write a request to book a room in {hotel} hotel, {city}, from {check_in} to {check_out}')
        }
    )

Invalid booking request with missing dates

for _ in tqdm(range(32)):
    city = random.choice(cities)
    hotel = random.choice(hotels)
    user_queries.append(
        {
            'available': len(hotels_df.query('city == @city & name == @hotel')) > 0,
            'query': generate_query(f'Write a request to book a room in {hotel} hotel, {city}')
        }
    )

Invalid booking request with missing hotel names

for _ in tqdm(range(32)):
    city = random.choice(cities)
    check_in, check_out = random_stay()
    user_queries.append(
        {
            'available': len(hotels_df.query('city == @city & name == @hotel')) > 0,
            'query': generate_query(f'Write a request to book a room in a hotel in {city} from {check_in} to {check_out}')
        }
    )

Invalid booking request with missing cities

for _ in tqdm(range(32)):
    hotel = random.choice(hotels)
    check_in, check_out = random_stay()
    user_queries.append(
        {
            'available': len(hotels_df.query('city == @city & name == @hotel')) > 0,
            'query': generate_query(f'Write a request to book a room in {hotel} hotel from {check_in} to {check_out}')
        }
    )

Irrelevant questions

def generate_irrelevant_query() -> str:
    response: ChatResponse = chat(
        'mistral',
        messages=[
            {'role': 'user', 'content': 'Write a question disrelated to hotel booking'}
        ],
        stream=False,
        options={'temperature': 1.0}
    )
    return response.message.content

for _ in tqdm(range(32)):
    user_queries.append(
        {
            'available': False,
            'query': generate_irrelevant_query()
        }
    )

df = pd.DataFrame.from_records(user_queries)
df

df.to_excel('dataset_raw.xlsx', index=False)

## Generate more booking requests
Random combinations of city and hotel names above resulted in mostly unavailable options. Need more valid valid queries

extra_queries = []
for _ in tqdm(range(128)):
    index = random.randint(0, len(hotels_df) - 1)
    hotel = hotels_df.at[index, 'name']
    city = hotels_df.at[index, 'city']
    check_in, check_out = random_stay()
    extra_queries.append(
        {
            'available': True,
            'query': generate_query(f'Write a request to book a room in {hotel} hotel, {city}, from {check_in} to {check_out}')
        }
    )

addendum = pd.DataFrame.from_records(extra_queries)

addendum

addendum.to_excel('dataset_raw_addendum.xlsx', index=False)
