==> 03_train_test_split.ipynb <==
import pandas as pd

df = pd.read_excel('dataset_final.xlsx')
df

train = df.sample(frac=0.8,random_state=42)
test = df.drop(train.index)

train.to_csv('train.csv', index=False)

test.to_csv('test.csv', index=False)
