==> 06_quantize_model.ipynb <==
import os, json

import torch
from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig
from dotenv import load_dotenv

%load_ext chime

load_dotenv()
HF_TOKEN = os.getenv('HF_TOKEN')

bnb_config = BitsAndBytesConfig(
    load_in_4bit=True,
    bnb_4bit_quant_type="nf4",
    bnb_4bit_compute_dtype=torch.float16,
    bnb_4bit_use_double_quant=True,
)

%%chime
model = AutoModelForCausalLM.from_pretrained(
    "/home/denis/Models/LLM/Mistral-7B-Instruct-v0.3/",
    quantization_config=bnb_config,
    device_map="auto",
    local_files_only=True,
    token=HF_TOKEN
)

tokenizer = AutoTokenizer.from_pretrained("mistralai/Mistral-7B-Instruct-v0.3")

if tokenizer.pad_token is None:
    tokenizer.pad_token = tokenizer.eos_token 

path = '/home/denis/Models/LLM/Mistral-7B-Instruct-v0.3_quantized_4bit/'

model.save_pretrained(path)

tokenizer.save_pretrained(path)

with open(os.path.join(path, "quant_config.json"), "w") as f:
    json.dump(bnb_config.to_dict(), f)
