==> 07_prepare_trainset.ipynb <==
# Prepare a train set to fine-tune the model

import pandas as pd
from tqdm import tqdm
tqdm.pandas()

from booking_baseline import BookingAgent, available_functions, system_message_1, system_message_2

booking_agent = BookingAgent(
    functions=available_functions,
    model_name='mistral',
    system_message_1=system_message_1,
    system_message_2=system_message_2
)

train_df = pd.read_csv('train.csv')
train_df

train_df[['llm_response_1', 'llm_response_2']] = train_df.progress_apply(lambda row: booking_agent(row['query']), axis=1, result_type='expand')

train_df

train_df.to_excel('train_raw.xlsx', index=False)

## Finalize the train set to fine-tune the 1st stage (tool call generation)
The dataset was checked and corrected manually.

import json, os

import pandas as pd
from datasets import Dataset, DatasetDict
from transformers import AutoTokenizer

from booking_baseline import system_message_1

train_df = pd.read_excel('train_first_response_only.xlsx')
train_df

model_path = '/home/denis/Models/LLM/Mistral-7B-Instruct-v0.3_quantized_4bit/'

Model's max number of tokens:

with open(os.path.join(model_path, 'config.json')) as f:
    model_config = json.load(f)
print(model_config)

model_config['max_position_embeddings']

Use a smaller number to save memory (checked that the maximum number of tokens in the train set is less than 2000)

max_tokens = 2048

tokenizer = AutoTokenizer.from_pretrained(model_path)

tools = [
    {
        "type": "function",
        "function": {
            "name": "get_hotels",
            "description": "Find hotels with rooms available for booking in a given city between given check-in and check-out dates",
            "parameters": {
                "type": "object",
                "properties": {
                    "city": {
                        "type": "string",
                        "description": "The city where available hotels should be found",
                    },
                    "checkin_date": {
                        "type": "string",
                        "description": "Date when the user wants to check-in",
                    },
                    "checkout_date": {
                        "type": "string",
                        "description": "Date when the user wants to check-out",
                    }
                },
                "required": ["city", "checkin_date", "checkout_date"],
            },
        },
    },
    {
        "type": "function",
        "function": {
            "name": "book_hotel",
            "description": "Book a room in the specified hotel from the provided check-in date to the provided check-out date",
            "parameters": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the hotel where the user wants a room to be booked",
                    },
                    "city": {
                        "type": "string",
                        "description": "The city where the hotel is located",
                    },
                    "checkin_date": {
                        "type": "string",
                        "description": "Date when the user wants to check-in",
                    },
                    "checkout_date": {
                        "type": "string",
                        "description": "Date when the user wants to check-out",
                    }
                },
                "required": ["name", "city", "checkin_date", "checkout_date"],
            },
        },
    }
]

ds = Dataset.from_pandas(train_df[['query', 'llm_response_1']], split="train", preserve_index=False)

def preprocess_function(examples):
    inputs = [
        f'''System:
{system_message_1}
Tools:
{json.dumps(tools, indent=2)}
User request:
{query}'''
        for query in examples["query"]
    ]
    targets = examples["llm_response_1"]
    model_inputs = tokenizer(
        inputs,
        padding='max_length',
        truncation=True,
        max_length=max_tokens
        
    )
    labels = tokenizer(
        targets,
        padding='max_length',
        truncation=True,
        max_length=max_tokens
    )
    model_inputs["labels"] = labels["input_ids"] 
    return model_inputs

tokenized_datasets = ds.map(preprocess_function, batched=True, remove_columns=ds.column_names)

tokenized_datasets.save_to_disk('./train_ds_1st_stage/')
