==> 08_finetune_1st_stage.ipynb <==
import os, json

from transformers import AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig
from peft import LoraConfig, get_peft_model
from transformers import TrainingArguments, Trainer, DataCollatorWithPadding, DataCollatorForLanguageModeling

from datasets import load_from_disk

%load_ext chime

quantized_model_path = '/home/denis/Models/LLM/Mistral-7B-Instruct-v0.3_quantized_4bit_compute_float32/'

with open(os.path.join(quantized_model_path, "quant_config.json")) as f:
    quant_config = json.load(f)

bnb_config = BitsAndBytesConfig(**quant_config)

model = AutoModelForCausalLM.from_pretrained(
    quantized_model_path,
    quantization_config=bnb_config,
    device_map="auto",
    local_files_only=True
)

peft_config = LoraConfig(
    r=8,
    lora_alpha=16,
    target_modules=["q_proj", "v_proj"],
    lora_dropout=0.05,
    bias="none",
    task_type="CAUSAL_LM"
)

model = get_peft_model(model, peft_config)
model.enable_input_require_grads() 

model.print_trainable_parameters()

print(model.config.quantization_config.to_dict())

tokenizer = AutoTokenizer.from_pretrained(quantized_model_path)

data_collator = DataCollatorWithPadding(tokenizer=tokenizer)

data_collator = DataCollatorForLanguageModeling(
    tokenizer=tokenizer,
    mlm=False 
)

training_args = TrainingArguments(
    output_dir='training_1st_stage',
    overwrite_output_dir=True,
    do_train=True,
    do_eval=False,
    eval_strategy='no',
    fp16=True,
    per_device_train_batch_size=2,
    gradient_accumulation_steps=2,
    learning_rate=2e-5,
    weight_decay=1e-4,
    prediction_loss_only=False,
    num_train_epochs=1,
    logging_strategy='steps',
    logging_steps=4,
    save_strategy='no',
    use_cpu=False,
    group_by_length=False,
    gradient_checkpointing=True,
    remove_unused_columns=False,
    label_names=["labels"]
)

ds = load_from_disk('train_ds_1st_stage')
ds

trainer = Trainer(
    model=model,
    args=training_args,
    data_collator=data_collator,
    train_dataset=ds,
    tokenizer=tokenizer,
)

%%chime
trainer.train()

trainer.save_model("./model/fine_tuned_1st_stage")

tokenizer.save_pretrained("./model/fine_tuned_1st_stage")
