==> 11_eval_1st_stage_finetune.ipynb <==
# Evaluate baseline results

import pandas as pd

results = pd.read_excel('finetune_1st_stage_test_results.xlsx')
results

Share of correct output from the first call to LLM:

results.correct_1.mean().item()

Share of correct output from the second call to LLM:

results.correct_2.mean().item()

## Results
Fist call to the LLM generates correct function calls or messages in 78% cases: no change after fine-tuning.

(Unexpectedly, the second call results improved by 6% to 66%, though the model still believes erroneously that many requested hotels are in Moscow.)
